WIP: cmd/compile: enable carry chain scheduling for AMD64
Feedback is welcome, as a PPC64 maintainer, this might not be
the desired or ideal implementation for AMD64. Likewise, the
priority changes may not be needed/desired on all architectures.

AMD64 arithmetic seems to universally clobber flags, this causes
entire carry chains to regenerate when compiling non-trivial
functions like those in crypto/elliptic/internal/fiat.

Unlike PPC64 where few instructions clobber the isolated carry
bit, the scheduler needs to minimize scheduling of unrelated ops
during the chain. To accomplish this, we bump the priorities on
the select ops and carry bit operations above the regular stuff
to ensure they are always chosen ahead of unrelated ops.

Likewise, the ssa rules need tweaked to ensure unrelated chains
are not merged, end of chains are easier to identify and avoid
folding which may cause overlapping chains.

This results in about a 2x improvement in the crypto/elliptic
benchmarks which use the go-native curves on AMD64.

Updates #40171

Change-Id: I8ba37356a565f66099498cf5781d50bc3e23ae8c