runtime: create async work queue to handle runtime triggers
Currently the runtime wants to wake the scavenger from within the
sweeper, which is called from the allocator, but can't call into the
allocator from there -- it could be re-entrant, leading to
self-deadlock! Instead, it relies on sysmon to notice that the goroutine
is awoken.

In the near future we're going to want to trigger GCs from the
allocator, or at least notify the runtime that one should be triggered
promptly. GCs need to be triggered from a regular goroutine for a number
of reasons, but in general it's because gcStart expects to be running on
a regular goroutine. Unfortunately, that promptness that we desire
prevents us from using sysmon for this purpose, because it may take up
to 10 ms for sysmon to notice (or more, if the application is idle).

We have this general problem now in the runtime of wanting to trigger
work items asynchronously, so this CL adds a queue for such work. The
runtime manages a worker goroutine that blocks until a work item comes
in, at which point any M in the scheduler should notice the work item
and execute it.

This change folds scavenger waking and periodic GCs into this new
system, in preparation for triggering GCs this way more generally.

For #48409.

Change-Id: I9731f3e709a4c9fb5923afc0421180de52a377ba