cmd/compile: convert merge to use appl. bal. trees for sharing
DO NOT REVIEW (yet)

This CL replaces a not-very-shared linear-sized set
representation with a much more shared representation.
For the annoying test program in question, it reduces
the heap size by 95%.  HOWEVER, for smaller versions
of this program at least, it takes longer.

However, for some programs build time is longer.

This also includes at least one bug fix for problems
uncovered while ensuring compatibility with what it
replaces.

Updates #51543.

Change-Id: Ie7a4c6ea460775faeed2b0378ab21ddffd15badc