cmd/compile: convert merge to use appl. bal. trees for sharing
DO NOT REVIEW (yet)

This CL replaces a not-very-shared linear-sized set
representation with a much more shared representation.
For the annoying test program in question, it reduces
the heap size by 95%.  HOWEVER, for smaller versions
of this program at least, it takes longer.

There's room for improvement.
There's also a test suite for the balanced binary trees,
not yet checked in.

Updates #51543.

Change-Id: Ie7a4c6ea460775faeed2b0378ab21ddffd15badc