go/types, types2: use declaring interface as receiver for interface methods
Functions that are methods are identified as such by the presence
of a non-nil receiver. For interface methods, the receiver is the
interface that declares the method. If the interface is given a
type name as in

        type T interface {
                m()
        }

before this CL, the method receiver type was the named type T. If
the interface is anonymous, the method receiver is the respective
interface literal.

With this change, the receiver type of interface methods is always
the respective interface literal.

Fixes #49906.

Change-Id: I0ecda1416347e17fb6fdd0b44cd8a64131b168f1