go/types, types2: better error message for invalid type parameter term
The spec says "In a union, a term cannot be a type parameter,...",
but it's really the type in a term (the term type) that cannot be
a type parameter. Also, for the spec's purposes, a single term is
still a union.

This CL changes the current error message from:

        "cannot use type parameter in typeset"

to

        "term type cannot be a type parameter"

which is more specific and matches the spec more closely.

Fixes #50420.

Change-Id: Id48503efc8416cabc03d5c40d8e64d5b3a7f078e