arena: have nil arenas allocate from the heap
DO NOT REVIEW
DO NOT SUBMIT

This makes the zero value of *arena.Arena useful. Arena consumers can
use the same code for heap and arena allocations. In particular, this
allows types to add a SetArena method and manage allocations through an
arena regardless of whether the method has been called.

For #51317

Change-Id: I8c3bbca6433f57f076976ae3906e6d11660420e5