types2: fix overlap test for union termlist
DO NOT REVIEW.
- needs go/types port

Per the spec, "the type sets of all non-interface terms must be
pairwise disjoint (the pairwise intersection of the type sets must
be empty)" in a union.

For the overlap test, the existing implementation casually mixed
syntactic union terms (which may have interface type) with type set
terms (which are normalized/expanded and must not have interface
type). As a consequence, in some cases the overlap test failed.

This CL carefully separates syntactic terms ("Term") from type set
terms ("term") by giving them explicitly different types so that
they cannot easily be converted from one to the other, and by using
explicit factory functions in all cases. It also changes the order
of their struct fields to limit the possibility of the use of a
composite literal where a factory function should be used.

Replace the overlap function with a proper method of termlists,
where it belongs.

Fixes #51607.

Change-Id: Ie3f7a16424b13e3d94c4d4e73e8979e82a20c9a1