go/types, types2: separate implementation of term from Term
The fix for issue #51607 was to not mix term and Term values
through careless conversions between them.

This change is a follow-up that carefully separates syntactic
terms ("Term") from type set terms ("term") by giving them
explicitly different types so that they cannot easily be
converted from one to the other, and by using explicit
factory functions in all cases.

Replace the overlap function with a proper method of termlists,
where it belongs.

For #51607.

Change-Id: Ie3f7a16424b13e3d94c4d4e73e8979e82a20c9a1