runtime: update inconsistent gcController stats more carefully
Currently the gcController stats are updated arbitrarily, but soon we're
going to be calculating an expression from them that is important does
not overflow. The most important thing to avoid is double-counting
situations when these values are updated together (namely, heapFree,
mappedReady, and totalAlloc-totalFree).

This change ensures double-counting is avoided between these values, and
leaves breadcrumbs as to why. A later CL will actually be enforcing
this.

For #48409.

Change-Id: Ife52caccc18cb826fb31f117709c4a66a4cba5ae