runtime: replace PI controller in pacer with simpler heuristic
This change replaces the PI controller in the pacer, which tends to be
bouncy and produce weird results like negative cons/mark ratios (which
is necessary to allow it to "hunt").

Instead, we use a simple heuristic of picking the median cons/mark
ratio we computed over the last 3 GC cycles.

This heuristic is much better at eliminating anomalies in the cons/mark
ratio due to noise, as it is akin to median filtering, which is
particularly good at 'despeckling.'

For #48409.

Change-Id: I5f1036057d1bbe58e98c0169b8b60f590835e65c