cmd/link: complete decoupling for newtype
Implement the interface dwarf.Type for dwsym, and use it in newtype1. Now, newtype1 is no related to any specific code with linker, it will moved to cmd/internal/dwarf next cl.

note:
1. Using dwsym as interface will cost a lot of time because of duplicate
function call. But in the future, the most of dwarf type info will be generated in compiler, these will not be a hotpoint. Or, maybe it can be optimizated by cache.
2. make Type.DefGoType and  Type.DefPtrTo as one function.

For #52209

Change-Id: I30165d0bb7722924ed1f306a2807fabf3eb4c22f