cmd/compile,go/build: allow std to use go:embed without io/fs
Some standard library packages can't easily import the embed package.
One such example is time/tzdata, as that results in an import cycle:

		time/tzdata -> embed -> io/fs -> context -> time -> time/tzdata

The import cycle is real, as the embed package uses fs.FS in its API.
However, it's very unfortunate in the case of time/tzdata,
because we really just want to embed a file as a string var.
We do not actually need to import the io/fs package at all.

A significant portion of standard library packages are in this position,
given that io/fs imports 27 packages in total.
There's really no good reason that those packages shouldn't be able to
use go:embed with string or []byte variables,
as that really only requires the toolchain to do the right thing.

An alternative approach that was considered was to relax the rule that
"embed" must be imported when compiling standard library packages.
However, that implementation was measurably slower: go/build now had to
scan all Go files in the standard library for "//go:build" comments,
whereas before it avoided that when the "embed" import was missing.
Imports are always at the top of the file and the import is rare,
so that shortcut actually made a significant difference.

Introducing a new package is slightly unfortunate, but it's very low
cost given how it's an empty package. We then teach go/build and the
compiler that importing it also allows using go:build directives.
Note that the package is in internal/reflectlite, so the alternative is
only available to the standard library and toolchain as per the internal
package rules.

Fixes #51463.

Change-Id: I233081276d5c4d718d183a84db68298130be0cc3