slices: use !Less in SortStable
In CL 371574, we replaced all !{{Less}} with {{GreaterOrEqual}} to fix a problem(handle NaNs when sorting float64 slice) in exp/slices:UnstableSort, but the change will also affect the exp/slices:StableSort.

This CL reverts all the replacements in exp/slices:StableSort and does not affect any codes in the sort package.

Change-Id: Idc225d480de3e2efef2add35c709ed880d1306cb