encoding/xml: fix 'unsupported type' error on interface{} attributes
When given interface value that is actually a type implementing
`encoding.TextMarshaler` or `xml.MarshalerAttr`, `marshalAttr` would return an
`unsupported type` error. The cause of this is that pointer and interface values
are dereferences after checking if the supported interfaces are implemented.

Solve this by moving the dereference of the pointer and interface values to the
start of the function, and update the test cases to test for this situation.

Fixes #48624

Change-Id: I1727ac0fd94bf9ea545d89a5fe3cea2496cb81f8
GitHub-Last-Rev: 16f22f78caa3c02f67fe34b6468040f58f0a788d
GitHub-Pull-Request: golang/go#48625