cmd/compile: synthesize types in compiler
Now, we can synthesize types in compiler. And all the dwarf type sym can be kept by relocs. subtype in linker can be removed now. There is some undifined relocs to runtime type in dwarf type.
These relocs are added by compiler, it don't know the corresponding runtime type will be defined or not. So skip them in relocsym.

For #52209

Change-Id: Ief7683f8d9033ebf0f682f582e2fffde63587bf0