cmd/compile,link: support generating dwarf info when -linkshared
Now, all the dwarf type can be generated by compiler, when using static link, every compiler unit generate its own dwarf type die. when using -linkshared, all the types current compile unit need can be collected. So we can generate corresponding dwarf type info for it and subtypes of it.
For synthesizetype, we also need to keep the type info of them when linkshared

For #52209

Change-Id: Ia70ba39708ce710c45d923d5b86dac91d7e9d616