runtime: add hook to register coverage-instrumented packages
DO NOT SUBMIT

Add support to the runtime for registering coverage-instrumented
packages, using a new hook that can be called from the init function
of an instrumented package. The hook records the meta-data symbol for
the package (chaining it onto a list), and returns a package ID to be
used to identify functions in the package. This new hook is not yet
called; that will be added in a subsequent patch. The list of
registered meta-data objects will be used (again in a future patch) as
part of coverage data file writing.

Special handling is required for packages like "runtime" or
"internal/cpu", where a function in the package may be executed before
the package "init" func runs. For these packages hard-code the package
ID, then make sure that the ID matches when the init func does run.

Change-Id: I6ca3ddf535197442a2603c6d7a0a9798b8496f40