reflect: add examples for (Slice|String)Header
go#40701 shows that users in the wild are prone to misusing them.
This is possibly due to the fact that the documentation does not
actually show the correct way in which they should be used.

Personally speaking, even if I am very well aware of the existence
of this potential footgun, I have to look up every time the correct
incanation to use, and as the docs currently do not include it I
have to chase down random blogposts or github issues to remember.

This CL adds the correct examples right in the docs, so that people
with bad memory like me can find it right away.

Change-Id: I3b75819409177b5a286c1e9861a2edb6fd1301b8