os/exec: clean up pipe-closing logic
Rename closeAfterStart and closeAfterWait to localPipes and
remotePipes respectively. That makes their contents clearer, and also
helps to clarify what should happen on error (when, for example, Wait
shouldn't be called at all).

Use a deferred call instead of individual calls to close pipe FDs
after Start. That helps to clarify the invariants around when they are
closed, and also makes the function a bit more robust for future
refactoring.

Also nil out the slices containing the pipe files so that they can be
collected earlier.

This CL is intended as a pure refactor in preparation for #50436.

Change-Id: I05d13fa91d539b95b84b2ba923c1733f9a6203e5