cmd/compile: modify test, compile, runtime for set LocalPkg.Path to -p flag
The modified test is for behavior that is undocumented,
internal, and not strictly correct, so I changed it.  No
other tests failed from this change.

The change to obj.go seems necessary to get the right amount
of symbol definitions on arm64/linux, in misc/cgo/testshared.

The change to runtime is to turn off now-detected violations
of nowritebarrierrec; except for js/wasm, these were cases
of panics that "should never happen", and that were already
there, we were just blissfully ignorant.

For js/wasm, we hope that the peculiar runtime environment
prevents the newly-detected bugs from ever going wrong, but,
again, the bug was already there.

The change to GOEXPERIMENT treatment is uncertain; it fixes
the bug and seems harmless. I'm still looking for why this
is NOT a bug before the prior CL.

Change-Id: Idbb3bc6fa4e41c22d9a125a025d0c2185e732fe0