crypto/tls,net/http: consolidate TLS event logging
The existing httptrace code that invokes the TLS-specific user-specified
callbacks during an HTTP request is spread throughout the net/http
package. The code in this patch fixes that and adds tracing functionality
to crypto/tls a la httptrace. Compatibility with the existing httptrace
API is maintained.

Part of a Fixes #52110. A full fix for #52110 will require a separate
change to x/net/http2.

Change-Id: Ia55544902aea71752940b8c7d4918b8599618afc